Move the schema-shaped chat tool helpers into @valbuild/shared/internal
The MCP tool registry needs the same patch builders and validation filter the
Studio's chat tools use, and it cannot have them: @valbuild/ui exports only
"." and "./server", so packages/server cannot reach spa/**. @valbuild/shared
is importable from both realms and depends on nothing but @valbuild/core and
zod, so that is where this logic belongs - none of it was ever React or DOM
dependent, it just happened to live next to the components that called it
first.

Moved, with their tests: aiSourceToolPatches, aiImageToolPatches, toolNames,
blockingValidationErrors, partitionValidationErrors (a dependency of the
last), and emptyOf. Pure renames plus one-line import rewrites at 15 call
sites - no logic changed, which is the point: this lands on its own so that
any regression in the live chat path is attributable to a move rather than
hidden inside the registry work that follows.

emptyOf's single date-fns call is inlined instead of adding a date library to
shared for one format string. The replacement formats the LOCAL date, matching
date-fns; toISOString().slice(0, 10) would have been UTC and handed anyone
west of Greenwich yesterday's date as their "today" default.

Verified: pnpm run -r typecheck, pnpm run lint, prettier --check, and the full
suite (2768 tests, 221 suites) all green, with the four moved test files
passing from their new location.
