+++ packages/shared/src/internal/aiTools/aiImageToolPatches.test.ts
