+++ packages/shared/src/internal/aiTools/aiSourceToolPatches.test.ts
