+++ packages/shared/src/internal/index.ts
@@ -16,3 +16,14 @@ export * from "./resolveSchemaSourceFixes";
 export * from "./getErrorMessageFromUnknownJson";
 export * from "./zod/SerializedSchema";
 export * from "./valCanvasProtocol";
+// Schema-shaped helpers that both realms need: the Studio's chat tools run in
+// the browser, and the MCP tool registry in @valbuild/server runs server-side,
+// but the patches they build and the errors they filter are the same logic.
+// Kept here rather than duplicated, since @valbuild/server cannot import from
+// @valbuild/ui's spa/**.
+export * from "./emptyOf";
+export * from "./aiTools/toolNames";
+export * from "./aiTools/aiImageToolPatches";
+export * from "./aiTools/aiSourceToolPatches";
+export * from "./validation/partitionValidationErrors";
+export * from "./validation/blockingValidationErrors";
