+++ packages/shared/src/internal/validation/partitionValidationErrors.test.ts
