+++ packages/shared/src/internal/validation/partitionValidationErrors.ts
