+++ packages/ui/spa/components/AddRecordPopover.tsx
@@ -1,7 +1,7 @@
 import { SourcePath, Internal, ModuleFilePath } from "@valbuild/core";
 import { JSONValue } from "@valbuild/core/patch";
 import { useState, useEffect, useCallback } from "react";
-import { emptyOf } from "./fields/emptyOf";
+import { emptyOf } from "@valbuild/shared/internal";
 import { Button } from "./designSystem/button";
 import { Input } from "./designSystem/input";
 import {
