+++ packages/ui/spa/stores/react/useValidationErrors.ts
@@ -1,5 +1,5 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
-import { filterBlockingValidationErrors } from "../../validation/blockingValidationErrors";
+import { filterBlockingValidationErrors } from "@valbuild/shared/internal";
 import {
   Internal,
   type ModuleFilePath,
